backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
pydantic>=1.8.2
starlette>=0.14.2
python-multipart>=0.0.5